make components smaller
to use everywhere each components
