+++ src/components/Users/Users.jsx
@@ -1,9 +1,13 @@
+import SingleUser from "./SingleUser";
+
 const Users = ({ UsersList }) => {
   return (
     <>
       <div className="flex relative">
         {UsersList.map((item) => {
-          return <img src={item.src} key={item.id} alt="" className={item.className} />;
+          return (
+            <SingleUser src={item.src} key={item.id} className={item.className}/>
+          )
         })}
       </div>
     </>
